requirements.txt
flask>=3.0
flask-cors>=4.0
pymongo>=4.6
PyJWT>=2.8
